refactor(loop-support): split capability lifecycle tests
